Refresh shiny app automatically

Refresh shiny app automatically

run_app_auto(port = 8181, host = getOption("shiny.host", "127.0.0.1"), ...)

Arguments

port
The TCP port that the application should listen on. If the port is not specified, and the shiny.port option is set (with options(shiny.port = XX)), then that port will be used. Otherwise, use a random port.

host
The IPv4 address that the application should listen on. Defaults to the shiny.host option, if set, or "127.0.0.1" if not. See Details.

...
Arguments passed on to shiny::runApp

appDir

The application to run. Should be one of the following:

A directory containing server.R, plus, either ui.R or a www directory that contains the file index.html.

A directory containing app.R.

An .R file containing a Shiny application, ending with an expression that produces a Shiny app object.

A list with ui and server components.

A Shiny app object created by shinyApp().

launch.browser

If true, the system's default web browser will be launched automatically after the app is started. Defaults to true in interactive sessions only. This value of this parameter can also be a function to call with the application's URL.

workerId

Can generally be ignored. Exists to help some editions of Shiny Server Pro route requests to the correct process.

quiet

Should Shiny status messages be shown? Defaults to FALSE.

display.mode

The mode in which to display the application. If set to the value "showcase", shows application code and metadata from a DESCRIPTION file in the application directory alongside the application. If set to "normal", displays the application normally. Defaults to "auto", which displays the application in the mode given in its DESCRIPTION file, if any.

test.mode

Should the application be launched in test mode? This is only used for recording or running automated tests. Defaults to the shiny.testmode option, or FALSE if the option is not set.